Conflicts:
	CMakeLists.txt
	src/icrar/leap-accelerate/CMakeLists.txt
	src/icrar/leap-accelerate/algorithm/cpu/CpuLeapCalibrator.cc
	src/icrar/leap-accelerate/core/stream_out_type.h
	src/icrar/leap-accelerate/math/cpu/matrix_invert.h

+++ src/icrar/leap-accelerate/algorithm/cpu/CpuLeapCalibrator.cc
@@ -85,7 +85,7 @@ namespace cpu
 
         profiling::timer calibration_timer;
 
-        size_t timesteps = (size_t)ms.GetNumTimesteps();
+        int timesteps = boost::numeric_cast<int>(ms.GetNumTimesteps());
         Range validatedSolutionInterval = solutionInterval.Evaluate(timesteps);
         std::vector<double> epochs = ms.GetEpochs();
 
@@ -116,7 +116,7 @@ namespace cpu
             const auto integration = Integration(
                     integrationNumber,
                     ms,
-                    solution * validatedSolutionInterval.interval * ms.GetNumBaselines(),
+                    boost::numeric_cast<int32_t>(solution * validatedSolutionInterval.interval * ms.GetNumBaselines()),
                     ms.GetNumChannels(),
                     validatedSolutionInterval.interval * ms.GetNumBaselines(),
                     ms.GetNumPols());
